requirements.txt
# 核心依赖
flask>=2.3.0
click>=8.1.0
pyyaml>=6.0
scikit-learn>=1.3.0
python-Levenshtein>=0.21.0
rich>=13.0.0
tabulate>=0.9.0

# 开发依赖
pytest>=7.4.0
pytest-cov>=4.1.0
flake8>=6.0.0
black>=23.0.0
